draft: google refactor
