--- src/core/Application/ITMO.Dev.ASAP.Application.Abstractions/Google/ISubjectCourseTableService.cs
@@ -1,6 +1,0 @@
-namespace ITMO.Dev.ASAP.Application.Abstractions.Google;
-
-public interface ISubjectCourseTableService
-{
-    Task<string> GetSubjectCourseTableId(Guid subjectCourseId, CancellationToken cancellationToken);
-}
