--- src/core/Application/ITMO.Dev.ASAP.Application.Abstractions/Google/Notifications/SubjectCourseGroupQueueUpdateNotification.cs
@@ -1,5 +1,0 @@
-using MediatR;
-
-namespace ITMO.Dev.ASAP.Application.Abstractions.Google.Notifications;
-
-public record SubjectCourseGroupQueueUpdateNotification(Guid SubjectCourseId, Guid GroupId) : INotification;
