--- src/core/Application/ITMO.Dev.ASAP.Application.Abstractions/Google/Notifications/SubjectCoursePointsUpdatedNotification.cs
@@ -1,5 +1,0 @@
-using MediatR;
-
-namespace ITMO.Dev.ASAP.Application.Abstractions.Google.Notifications;
-
-public record SubjectCoursePointsUpdatedNotification(Guid SubjectCourseId) : INotification;
