--- src/core/Application/ITMO.Dev.ASAP.Application.Abstractions/Google/Sheets/ISheet.cs
@@ -1,6 +1,0 @@
-namespace ITMO.Dev.ASAP.Application.Abstractions.Google.Sheets;
-
-public interface ISheet<in TModel>
-{
-    Task UpdateAsync(string spreadsheetId, TModel model, CancellationToken token);
-}
