+++ src/core/Application/ITMO.Dev.ASAP.Application.Abstractions/ITMO.Dev.ASAP.Application.Abstractions.csproj
@@ -11,6 +11,8 @@
     </ItemGroup>
 
     <ItemGroup>
+        <Folder Include="Google\Notifications\" />
+        <Folder Include="Google\Sheets\" />
         <Folder Include="Submissions\Models" />
     </ItemGroup>
