+++ src/core/Application/ITMO.Dev.ASAP.Application.Contracts/Study/Queues/Notifications/QueueUpdated.cs
@@ -3,7 +3,7 @@ using MediatR;
 
 namespace ITMO.Dev.ASAP.Application.Contracts.Study.Queues.Notifications;
 
-public class QueueUpdated
+public static class QueueUpdated
 {
     public record Notification(Guid SubjectCourseId, Guid GroupId, SubmissionsQueueDto SubmissionsQueue) : INotification;
 }
