+++ src/core/Application/ITMO.Dev.ASAP.Application.Contracts/Study/SubjectCourseGroups/Notifications/SubjectCourseGroupQueueOutdated.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace ITMO.Dev.ASAP.Application.Contracts.Study.SubjectCourseGroups.Notifications;
+
+public static class SubjectCourseGroupQueueOutdated
+{
+    public record Notification(Guid SubjectCourseId, Guid GroupId) : INotification;
+}
