+++ src/core/Application/ITMO.Dev.ASAP.Application.Contracts/Study/SubjectCourses/Notifications/SubjectCoursePointsOutdated.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace ITMO.Dev.ASAP.Application.Contracts.Study.SubjectCourses.Notifications;
+
+public static class SubjectCoursePointsOutdated
+{
+    public record Notification(Guid SubjectCourseId) : INotification;
+}
