--- src/core/Application/ITMO.Dev.ASAP.Application.Google/Dummy/Services/DummySubjectCourseTableService.cs
@@ -1,11 +1,0 @@
-using ITMO.Dev.ASAP.Application.Abstractions.Google;
-
-namespace ITMO.Dev.ASAP.Application.Google.Dummy.Services;
-
-public class DummySubjectCourseTableService : ISubjectCourseTableService
-{
-    public Task<string> GetSubjectCourseTableId(Guid subjectCourseId, CancellationToken cancellationToken)
-    {
-        return Task.FromResult(string.Empty);
-    }
-}
