+++ src/core/Application/ITMO.Dev.ASAP.Application.Handlers/Study/SubjectCourses/CreateSubjectCourseHandler.cs
@@ -1,3 +1,5 @@
+using ITMO.Dev.ASAP.Application.Contracts.Study.SubjectCourses.Notifications;
+using ITMO.Dev.ASAP.Application.Dto.SubjectCourses;
 using ITMO.Dev.ASAP.DataAccess.Abstractions;
 using ITMO.Dev.ASAP.DataAccess.Abstractions.Extensions;
 using ITMO.Dev.ASAP.Domain.Study;
@@ -11,10 +13,12 @@ namespace ITMO.Dev.ASAP.Application.Handlers.Study.SubjectCourses;
 internal class CreateSubjectCourseHandler : IRequestHandler<Command, Response>
 {
     private readonly IDatabaseContext _context;
+    private readonly IPublisher _publisher;
 
-    public CreateSubjectCourseHandler(IDatabaseContext context)
+    public CreateSubjectCourseHandler(IDatabaseContext context, IPublisher publisher)
     {
         _context = context;
+        _publisher = publisher;
     }
 
     public async Task<Response> Handle(Command request, CancellationToken cancellationToken)
@@ -31,6 +35,11 @@ internal class CreateSubjectCourseHandler : IRequestHandler<Command, Response>
         _context.SubjectCourses.Add(subjectCourse);
         await _context.SaveChangesAsync(cancellationToken);
 
-        return new Response(subjectCourse.ToDto());
+        SubjectCourseDto dto = subjectCourse.ToDto();
+
+        var notification = new SubjectCourseCreated.Notification(dto);
+        await _publisher.Publish(notification, default);
+
+        return new Response(dto);
     }
 }
