+++ src/core/Infrastructure/Integration/ITMO.Dev.ASAP.Integration.Google/ITMO.Dev.ASAP.Integration.Google.csproj
@@ -1,15 +1,19 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <ItemGroup>
-        <PackageReference Include="FluentSpreadsheets.GoogleSheets"/>
-        <PackageReference Include="Google.Apis.Sheets.v4"/>
-        <PackageReference Include="Google.Apis.Drive.v3"/>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection"/>
-        <PackageReference Include="Microsoft.Extensions.Hosting"/>
+        <PackageReference Include="FluentSpreadsheets.GoogleSheets" />
+        <PackageReference Include="Google.Apis.Sheets.v4" />
+        <PackageReference Include="Google.Apis.Drive.v3" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
+        <PackageReference Include="Microsoft.Extensions.Hosting" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\..\Application\ITMO.Dev.ASAP.Application.Abstractions\ITMO.Dev.ASAP.Application.Abstractions.csproj"/>
+        <ProjectReference Include="..\..\..\Application\ITMO.Dev.ASAP.Application.Abstractions\ITMO.Dev.ASAP.Application.Abstractions.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Sheets\" />
     </ItemGroup>
 
 </Project>
