+++ src/google/ITMO.Dev.ASAP.Google.Application.Abstractions/ISheet.cs
@@ -1,0 +1,6 @@
+namespace ITMO.Dev.ASAP.Google.Application.Abstractions;
+
+public interface ISheet<in TModel>
+{
+    Task UpdateAsync(string spreadsheetId, TModel model, CancellationToken token);
+}
