+++ src/google/ITMO.Dev.ASAP.Google.Application.DataAccess/ITMO.Dev.ASAP.Google.Application.DataAccess.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Domain\ITMO.Dev.ASAP.Google.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
