+++ src/google/ITMO.Dev.ASAP.Google.Application.Handlers/ITMO.Dev.ASAP.Google.Application.Handlers.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Application.Abstractions\ITMO.Dev.ASAP.Google.Application.Abstractions.csproj" />
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Application.Contracts\ITMO.Dev.ASAP.Google.Application.Contracts.csproj" />
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Application.DataAccess\ITMO.Dev.ASAP.Google.Application.DataAccess.csproj" />
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Application.Spreadsheets\ITMO.Dev.ASAP.Google.Application.Spreadsheets.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="MediatR" />
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
+    </ItemGroup>
+
+</Project>
