+++ src/google/ITMO.Dev.ASAP.Google.Application.Spreadsheets/ITMO.Dev.ASAP.Google.Application.Spreadsheets.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+</Project>
