+++ src/google/ITMO.Dev.ASAP.Google.Application.Spreadsheets/Models/SheetId.cs
@@ -1,0 +1,3 @@
+namespace ITMO.Dev.ASAP.Google.Application.Spreadsheets.Models;
+
+public record struct SheetId(int Value);
