+++ src/google/ITMO.Dev.ASAP.Google.Application.Spreadsheets/Services/ISpreadsheetService.cs
@@ -1,0 +1,8 @@
+using ITMO.Dev.ASAP.Google.Application.Spreadsheets.Models;
+
+namespace ITMO.Dev.ASAP.Google.Application.Spreadsheets.Services;
+
+public interface ISpreadsheetService
+{
+    Task<SpreadsheetCreateResult> CreateSpreadsheetAsync(string title, CancellationToken token);
+}
