+++ src/google/ITMO.Dev.ASAP.Google.Application/ITMO.Dev.ASAP.Google.Application.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\core\Presentation\ITMO.Dev.ASAP.Presentation.Contracts\ITMO.Dev.ASAP.Presentation.Contracts.csproj" />
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Application.Abstractions\ITMO.Dev.ASAP.Google.Application.Abstractions.csproj" />
+      <ProjectReference Include="..\ITMO.Dev.ASAP.Google.Application.DataAccess\ITMO.Dev.ASAP.Google.Application.DataAccess.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="FluentSpreadsheets" />
+      <PackageReference Include="FluentSpreadsheets.GoogleSheets" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
+    </ItemGroup>
+
+</Project>
