+++ src/google/ITMO.Dev.ASAP.Google.DataAccess/Repositories/SubjectCourseRepository.cs
@@ -1,0 +1,3 @@
+namespace ITMO.Dev.ASAP.Google.DataAccess.Repositories;
+
+public class SubjectCourseRepository { }
